Link C++ standard library for systems that need it.
Some systems like Linux need to link the C++ standard library. If Recast or DetourTileCache are used, this was done automatically, but if you only use Detour or DetourCrowd, the C++ standard library isn't linked. This can cause linking to fail in some cases.

With this change, we link the corresponding standard library for most systems.